Permissions

+++ src/resolvers/Mutation.js
@@ -1,14 +1,14 @@
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const { randomBytes } = require('crypto');
-const { promisify } = require('util');
-const { transport, makeANiceEmail } = require('../mail');
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+const { randomBytes } = require("crypto");
+const { promisify } = require("util");
+const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: Check if they are logged in
     if (!ctx.request.userId) {
-      throw new Error('You must be logged in to do that');
+      throw new Error("You must be logged in to do that");
     }
 
     const item = await ctx.db.mutation.createItem(
@@ -17,19 +17,18 @@ const Mutations = {
           // This is how we create a realtionship between user and item in PRISMA
           user: {
             connect: {
-              id: ctx.request.userId,
-            },
+              id: ctx.request.userId
+            }
           },
-          ...args,
-        },
+          ...args
+        }
       },
-      info,
+      info
     );
 
     return item;
   },
   updateItem(parent, args, ctx, info) {
-    // first take a copy of the updates
     const updates = { ...args };
     delete updates.id;
 
@@ -37,72 +36,67 @@ const Mutations = {
       {
         data: updates,
         where: {
-          id: args.id,
-        },
+          id: args.id
+        }
       },
-      info,
+      info
     );
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
-    // 2. Check if they own that item, or have the permissions
-    // TODO
-    // 3. Delete it
-    console.log(info);
-    return ctx.db.mutation.deleteItem({ where }, info);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+    const ownsItem = (item.user.id = ctx.request.userId);
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+    const isAllowedToDelete = ownsItem || hasPermissions;
+
+    if (isAllowedToDelete) {
+      return ctx.db.mutation.deleteItem({ where }, info);
+    } else {
+      throw new Error("You don't have permissions to do that!");
+    }
   },
   async signup(parent, args, ctx, info) {
-    // lowercase their email
     args.email = args.email.toLowerCase();
-    // hash their password
     const password = await bcrypt.hash(args.password, 10);
-    // create the user in the database
     const user = await ctx.db.mutation.createUser(
       {
         data: {
           ...args,
           password,
-          permissions: { set: ['USER'] },
-        },
+          permissions: { set: ["USER"] }
+        }
       },
-      info,
+      info
     );
-    // create the JWT token for them
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // We set the jwt as a cookie on the response
-    ctx.response.cookie('token', token, {
+    const oneYear = 1000 * 60 * 60 * 24 * 365;
+    ctx.response.cookie("token", token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365, // 1 year cookie
+      maxAge: oneYear
     });
-    // finalllly we return the user to the browser
     return user;
   },
   async signin(parent, { email, password }, ctx, info) {
-    // 1. check if there is a user with that email
     const user = await ctx.db.query.user({ where: { email } });
     if (!user) {
       throw new Error(`No such user found for email ${email}`);
     }
-    // 2. check if their password is correct
     const valid = await bcrypt.compare(password, user.password);
     if (!valid) {
-      throw new Error('Invalid password!');
+      throw new Error("Invalid password!");
     }
-    // 3. generate the JWT token
     const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
-    // 4. Set the cookie with the token
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365,
+      maxAge: 1000 * 60 * 60 * 24 * 365
     });
-    // 5. Return the user
     return user;
   },
   signout(parent, args, ctx, info) {
-    ctx.response.clearCookie('token');
-    return { message: 'Goodbye' };
+    ctx.response.clearCookie("token");
+    return { message: "Goodbye" };
   },
   async requestReset(parent, args, ctx, info) {
     // 1. Check if this is a real user
@@ -112,27 +106,26 @@ const Mutations = {
     }
 
     // 2. Set a reset tokn and expiry on that user
-    const resetToken = (await promisify(randomBytes)(20)).toString('hex');
+    const resetToken = (await promisify(randomBytes)(20)).toString("hex");
     const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
     const res = await ctx.db.mutation.updateUser({
       where: { email: args.email },
-      data: { resetToken, resetTokenExpiry },
+      data: { resetToken, resetTokenExpiry }
     });
 
     // 3. Email them that reset token
     const mailRes = await transport.sendMail({
-      from: 'sartonon@hotmail.com',
+      from: "sartonon@hotmail.com",
       to: user.email,
-      subject: 'Your Password',
+      subject: "Your Password",
       html: makeANiceEmail(
         `Your password reset token is here! \n\n <a href="${
           process.env.FRONTEND_URL
-        }/reset?resetToken=${resetToken}">Click here to Reset</a>`,
-      ),
+        }/reset?resetToken=${resetToken}">Click here to Reset</a>`
+      )
     });
 
-    // 4. Return the message
-    return { message: 'Thanks!' };
+    return { message: "Thanks!" };
   },
   async resetPassword(parent, args, ctx, info) {
     // 1. check if the password match
@@ -144,11 +137,11 @@ const Mutations = {
     const [user] = await ctx.db.query.users({
       where: {
         resetToken: args.resetToken,
-        resetTokenExpiry_gte: Date.now() - 3600000,
-      },
+        resetTokenExpiry_gte: Date.now() - 3600000
+      }
     });
     if (!user) {
-      throw new Error('This token is either invalid or expired');
+      throw new Error("This token is either invalid or expired");
     }
     // 4. Hash their new password
     const password = await bcrypt.hash(args.password, 10);
@@ -158,19 +151,46 @@ const Mutations = {
       data: {
         password,
         resetToken: null,
-        resetTokenExpiry: null,
-      },
+        resetTokenExpiry: null
+      }
     });
     // 6. Generate JWY
     const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
     // 7. Set the JWT cookie
-    ctx.response.cookie('token', token, {
+    ctx.response.cookie("token", token, {
       httpOnly: true,
-      maxAge: 1000 * 60 * 60 * 24 * 365,
+      maxAge: 1000 * 60 * 60 * 24 * 365
     });
     // 8. return new user
     return updatedUser;
   },
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in!");
+    }
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId
+        }
+      },
+      info
+    );
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
+  }
 };
 
 module.exports = Mutations;
